Extract transaction manager in separate module
